Tests pass but a mess and typing needs to be resolved.
